ArtifactID: 891483 Comment:
Fixed issues with compiling on IA64.

Note:  IPMI is breaking the build now due to IA64 issues (on ia64).  A fix coming soon

+++ snmp/snmp_util.h
@@ -61,7 +61,8 @@
 struct snmp_value {
         u_char type;
         char string[MAX_ASN_STR_LEN];
-        unsigned int str_len;
+        size_t str_len;
+        //unsigned int str_len;
         long integer;
 };
